refactor: 환경변수 스키마 파일명 케밥 케이스로 변경
